Move DeviceContext from MyApp to contexts/device
