+++ jsconfig.json
@@ -5,7 +5,8 @@
       "@/components/*": ["components/*"],
       "@/lib/*": ["lib/*"],
       "@/layouts/*": ["layouts/*"],
-      "@/styles/*": ["styles/*"]
+      "@/styles/*": ["styles/*"],
+      "@/contexts/*": ["contexts/*"]
     }
   }
 }
