+++ pages/_app.js
@@ -1,22 +1,13 @@
-import React, {useEffect, createContext, useState} from 'react'
+import React, {useEffect} from 'react'
 import PropTypes from 'prop-types'
 import Head from 'next/head'
 
+import {DeviceContextProvider} from '@/contexts/device'
+
 const PIWIK_URL = process.env.NEXT_PUBLIC_PIWIK_URL
 const PIWIK_SITE_ID = process.env.NEXT_PUBLIC_PIWIK_SITE_ID
-const MOBILE_WIDTH = 900
-
-export const DeviceContext = createContext()
 
 function MyApp({Component, pageProps, isSafariBrowser}) {
-  const [viewHeight, setViewHeight] = useState('100vh')
-  const [isMobileDevice, setIsMobileDevice] = useState(false)
-
-  const handleResize = () => {
-    setViewHeight(`${window.innerWidth}px`)
-    setIsMobileDevice(window.innerWidth < MOBILE_WIDTH)
-  }
-
   const logPageView = () => {
     if (window.Piwik) {
       const tracker = window.Piwik.getTracker(`${PIWIK_URL}/piwik.php`, PIWIK_SITE_ID)
@@ -33,28 +24,14 @@ function MyApp({Component, pageProps, isSafariBrowser}) {
     }, 400)
   })
 
-  useEffect(() => {
-    window.addEventListener('resize', handleResize)
-    handleResize()
-
-    return () => {
-      window.removeEventListener('resize', handleResize)
-    }
-  }, [])
-
   return (
     <>
       <Head>
         <meta name='viewport' content='width=device-width, initial-scale=1' />
       </Head>
-      <DeviceContext.Provider value={{
-        viewHeight,
-        isMobileDevice,
-        isSafariBrowser
-      }}
-      >
+      <DeviceContextProvider isSafariBrowser={isSafariBrowser}>
         <Component {...pageProps} />
-      </DeviceContext.Provider>
+      </DeviceContextProvider>
       <style global jsx>{`
         body,
         html,
